+++ lib/rubocop/cop/layout/multiline_operation_indentation.rb
@@ -61,6 +61,7 @@ module RuboCop
         private
 
         def relevant_node?(node)
+          return false if node.unary_operation?
           !node.loc.dot # Don't check method calls with dot operator.
         end
 
@@ -72,7 +73,6 @@ module RuboCop
 
         def offending_range(node, lhs, rhs, given_style)
           return false unless begins_its_line?(rhs)
-          return false if lhs.first_line == rhs.line # Needed for unary op.
           return false if not_for_this_cop?(node)
 
           correct_column = if should_align?(node, rhs, given_style)